Remove deprecated API documentation for advanced user, computer, and mobile device searches
