+++ docs/generate/export/advancedcomputersearches.md
