+++ docs/generate/export/advancedmobiledevicesearches.md
